// src/services/types/board-page-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const boardPageResponseTypeEnum = {NOTICE: "NOTICE", BUILD_NOTE: "BUILD_NOTE"} as const;

export type BoardPageResponseTypeEnumKey = (typeof boardPageResponseTypeEnum)[keyof typeof boardPageResponseTypeEnum];

export type BoardPageResponse = {
	/**
	 * @description 게시글 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 게시판 유형
	 * @type string
	 */
	type: BoardPageResponseTypeEnumKey;
	/**
	 * @description 등급
	 * @type string
	 */
	grade: string;
	/**
	 * @description 제목
	 * @type string
	 */
	title: string;
	/**
	 * @description 내용
	 * @type string | undefined
	 */
	content?: string;
	/**
	 * @description 작성자 아이디
	 * @type integer, int64
	 */
	creatorId: number;
	/**
	 * @description 작성자 이름
	 * @type string | undefined
	 */
	creatorName?: string;
	/**
	 * @description 생성 일시
	 * @type string, date-time
	 */
	createdAt: string;
};